Fix dev onboarding errors. Add simple "repl"
There were errors that were preventing development. These errors came
up when there was a missing user, or a package could not be found. This
commit takes care of the errors in the simple case of a developer
starting from scratch.

During user signup, this commit skips the Stripe integration in a
non-production environment. This will allow easier development.

Lastly, it adds a simple http client "REPL" for sending commands to
Ava. It is located in cmd/avarepl/main.go.

Signed-off-by: Evan Tann <egtann@gmail.com>