Implements dynamic rpc networking
The existing framework for assigning ports/booting package RPC servers
was too brittle. This implemented a system where all rpc ports are
dynamically assigned by the OS.

Passing ":0" as the addr to a net.Listen() call causes the operating
system to allocate a random, available port. We use this method for all
RPC servers.

First, we boot the core ava RPC server and get the address. Then, we
pass that address to each package as a command line argument. Each
package boots it's own RPC server, and passes the dynamic address to the
core RPC server when it registers itself.

Signed-off-by: Evan Tann <egtann@gmail.com>