Refactor front-end
Changed mithril.route.mode to "pathname". This avoid the "?" in urls. This
required setting the root echo.Get route from "/" to "/*". Since echo will
match explicit routes first, there is no "route masking" issues

Added live-reload functionality. All this required was a added a simple script
to our index.html template and adding a couple of "cmd" routes to trigger the
reload from the server. Triggering a reload is as simple as calling "curl
localhost:4200/_/cmd/reload".  Along with this is a simple "fswatcher.sh"
script (in the "cmd/" dir) that listens for changes in the assets dir, rebuilds
the public dir from scratch, and calls reload for and open browsers. Note that
the first step in a rebuild is to delete all content from the public dir. This
is done to ensure clean and up to date builds. The script uses the fswatch
application but could be modified to use any other file system watcher. Note:
the live-reload/cmd functionality is only added in development builds

Lastly, the mithril component system has been standardized as follows:
- Each component is in a single file
- Components are wrapped in an IIFE
- Components only have a "controller" and "view function". Nothing else
        Note: some components still have "viewFull" functions that remain to be
        factored out
- Components begin with an uppercase letter

Misc changes:
- routes.js has been moved to app.js
- The google_oauth code is now in app.js
- app.js also contains "helper" functions that are available globally. By
  convention, these will start with a lowercase letter

Signed-off-by: Evan Tann <egtann@gmail.com>