Small markup change to install instructions
This again may be OS X-specific, but running a command like:
`cmd/setup.sh` won't properly export variables.

I'm supposing this is because running it like that will launch a
sub-shell, which doesn't propagate up the environment variables to
the parent shell.

So, a small minor markup change to run the setup script in the current
shell will fix the EXPORTs (and avoid needing to open a new shell).