Return 500s in the event of errors on signup
If the form validation goes through but there's an error at the database
level, or with creating the CSRF token, then return a 500.

We shouldn't allow the user to proceed, but it's not necessarily the
case that they've entered bad data (e.g. pg server being down would also
cause a failure).